requirements.txt
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.9.1
bcrypt==4.3.0
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
Django==5.2.5
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
ecdsa==0.19.1
fastapi==0.104.1
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.5.0
pydantic_core==2.14.1
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.6
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
sqlparse==0.5.3
starlette==0.27.0
typing_extensions==4.14.1
tzdata==2025.2
uvicorn==0.24.0
watchfiles==1.1.0
websockets==15.0.1
